moved covid utils
